refactor(api): move organizationAdministrationController.getOrganizationDetails to src/team
