+++ api/src/organizational-entities/application/organization/organization.admin.controller.js
@@ -15,6 +15,13 @@ const addOrganizationFeatureInBatch = async function (request, h) {
   return h.response().code(204);
 };
 
+const getOrganizationDetails = async function (request, h, dependencies = { organizationForAdminSerializer }) {
+  const organizationId = request.params.id;
+
+  const organizationDetails = await usecases.getOrganizationDetails({ organizationId });
+  return dependencies.organizationForAdminSerializer.serialize(organizationDetails);
+};
+
 const updateOrganizationsInBatch = async function (request, h) {
   await usecases.updateOrganizationsInBatch({ filePath: request.payload.path });
   return h.response().code(204);
@@ -23,5 +30,6 @@ const updateOrganizationsInBatch = async function (request, h) {
 export const organizationAdminController = {
   attachChildOrganization,
   addOrganizationFeatureInBatch,
+  getOrganizationDetails,
   updateOrganizationsInBatch,
 };
