+++ api/src/organizational-entities/domain/usecases/index.js
@@ -1,6 +1,7 @@
 import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 
+import * as schoolRepository from '../../../school/infrastructure/repositories/school-repository.js';
 import { injectDependencies } from '../../../shared/infrastructure/utils/dependency-injection.js';
 import { importNamedExportsFromDirectory } from '../../../shared/infrastructure/utils/import-named-exports-from-directory.js';
 import * as organizationFeatureRepository from '../../infrastructure/repositories/organization-feature-repository.js';
@@ -11,11 +12,13 @@ const path = dirname(fileURLToPath(import.meta.url));
 /**
  * @typedef {import ('../../infrastructure/repositories/organization-feature-repository.js')} OrganizationFeatureRepository
  * @typedef {import ('../../infrastructure/repositories/organization-for-admin.repository.js')} OrganizationForAdminRepository
+ * @typedef {import ('../../../school/infrastructure/repositories/school-repository.js')} SchoolRepository
  */
 
 const repositories = {
   organizationForAdminRepository,
   organizationFeatureRepository,
+  schoolRepository,
 };
 
 const dependencies = Object.assign({}, repositories);
@@ -28,6 +31,7 @@ const usecasesWithoutInjectedDependencies = {
  * @typedef OrganizationalEntitiesUsecases
  * @property {addOrganizationFeatureInBatch} addOrganizationFeatureInBatch
  * @property {attachChildOrganizationToOrganization} attachChildOrganizationToOrganization
+ * @property {getOrganizationDetails} getOrganizationDetails
  * @property {updateOrganizationsInBatch} updateOrganizationsInBatch
  */
